packages/seclea-ai/seclea_ai-1.1.0.tar.gz/seclea_ai-1.1.0/seclea_ai/lib/seclea_utils/dataset_management/dataset_utils.py
import pandas as pd
from pandas import DataFrame


def column_hash(dataset: DataFrame) -> int:
    """Invariant to ordering of columns/rows"""
    total = 0
    for col in dataset.columns:
        total += int(pd.util.hash_pandas_object(dataset[col]).sum())
    return total


def dataset_hash(dataset, project_id: str) -> str:
    return str(hash(column_hash(dataset) + hash(project_id)))


def get_dataset_project_hash(dataset, project) -> str:
    return str(hash(bytes(dataset)) + hash(project))


def get_dataset_hash(dataset) -> str:
    return str(hash(bytes(dataset)))


def get_project_hash(project) -> str:
    return str(hash(project))
